version works for AlgebraicABMs.jl as dependency of StateCharts.jl
This is the final version having StateCharts depending on AlgebraicABMs. And the example of with networks works good. The Infectious no network (simple model only with statechart) model is not correct, because it is not combined with the "basis" part of AlgebraicABMs.jl yet (having subpart of L match to the world)